requirements.txt
streamlit
pandas
alpaca-trade-api
yfinance
plotly
requests
matplotlib
ta
python-dotenv
